fix firestation?stationNumber station type (Long)

+++ src/main/java/com/openclassrooms/safetynets/dto/AddressDTO.java
@@ -9,7 +9,7 @@ public class AddressDTO {
 
 	String address;
 
-	Integer firestationNumber;
+	Long firestationNumber;
 
 	List<PersonWithAllergiesDTO> persons;
 }
